Changed Fixnum internal from int32_t to mrbc_int. Changed Float internal from double to mrbc_float.

+++ src/c_math.c
@@ -28,7 +28,7 @@ static double to_double( const mrb_value *v )
 {
   switch( v->tt ) {
   case MRB_TT_FIXNUM:	return (double)v->i;
-  case MRB_TT_FLOAT:	return v->d;
+  case MRB_TT_FLOAT:	return (double)v->d;
   default:		return 0;	// TypeError. raise?
   }
 }
